+++ src/lib.rs
@@ -7,17 +7,17 @@
 #[macro_use]
 extern crate failure;
 #[macro_use]
-extern crate lazy_static;
-#[macro_use]
 extern crate log;
-#[macro_use]
-extern crate paperclip_macros;
-#[macro_use]
-extern crate serde_derive;
 
 mod error;
 #[cfg(feature = "v2")]
 pub mod v2;
 
 pub use error::{PaperClipError, PaperClipResult};
-pub use paperclip_macros::api_v2_schema;
+#[cfg(feature = "v2")]
+pub use paperclip_macros::api_v2_schema_struct as api_v2_schema;
+
+#[cfg(feature = "actix")]
+pub mod actix {
+    pub use paperclip_actix::*;
+}